+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # ChangeLog
 
-* turns out that not all REI tooltips are actually tooltips.
+* fix CTM blocks rotating the model when CTM is applied
+* fix client not having a valid registry access on a server
+* fix memory leak in DummyWorld caused by the client level never being discarded.